+++ NEWS.md
@@ -1,3 +1,33 @@
+# manymome 0.1.11.2
+
+## New Features
+
+### Can report standard errors (if requested)
+
+- All major print methods of effects support
+  printing standard errors (setting
+  `se = TRUE`). They are simply the
+  standard deviations of the bootstrap
+  estimates (if bootstrap confidence
+  intervals are requested) or simulated
+  estimates (if Monte Carlo confidence
+  intervals are requested). They should
+  be interpreted with cautions because
+  the sampling distribution of the
+  effect estimates may not be symmetric. (0.1.11.2)
+
+## Miscellaneous
+
+- Customized `linters`. (0.1.11.1)
+
+## Bug Fixes
+
+- *P*-value were not computed when
+  mathematical operations are conducted
+  on effects using `+` and `-` before
+  version 0.1.11.2. This
+  has been fixed. (0.1.11.2)
+
 # manymome 0.1.10
 
 ## New Features